\name{isColorOut}
\alias{isColorOut}
\title{Check if colorizing of R output is enabled}
\description{
  Function to check if colorizing of R output is enabled.
}
\usage{
isColorOut()
}

\value{
  \code{TRUE} if the colorizing of R output is enabled and \code{FALSE}
  otherwise.
}

\author{Neal Fultz and Jakson Aquino \email{jalvesaq@gmail.com}}

