\name{setZero}
\alias{setZero}
\title{Enable colorizing of numbers near zero}
\description{
  Enable colorizing of numbers near zero
}
\usage{
setZero(z = 1e-12)
}

\arguments{
  \item{z}{The limit below which a value is considered close to 0.}
}

\value{
  NULL.
}

\seealso{\code{\link{unsetZero}}}

\author{Jakson A. Aquino \email{jalvesaq@gmail.com}}

