\name{setOutputColors256}
\alias{setOutputColors256}
\title{Deprecated Function}
\description{
  This function is deprecated. Use \link{setOutputColors} instead.
}
\usage{
setOutputColors256(...)
}
\arguments{
  \item{...}{Arguments to be passed to \link{setOutputColors}.}
}
\value{
  NULL.
}

\author{Jakson A. Aquino \email{jalvesaq@gmail.com}}
